backend/generator/requirements.txt
fastapi>=0.100
uvicorn>=0.20
google-cloud-aiplatform>=1.38
google-generativeai>=0.3
python-dotenv>=1.0
pillow>=9.0.0
supabase>=2.0.0
google-cloud-storage>=2.16.0
google-cloud-secret-manager>=2.16.0